+++ jupyterlab/index.js
@@ -21,7 +21,7 @@ var app = new phosphide.Application({
     require('jupyterlab/lib/landing/plugin').landingExtension,
     require('jupyterlab/lib/main/plugin').mainExtension,
     require('jupyterlab/lib/mainmenu/plugin').mainMenuExtension,
-    require('jupyterlab/lib/mdrenderwidget/plugin').markdownHandlerExtension,
+    require('jupyterlab/lib/markdownwidget/plugin').markdownHandlerExtension,
     require('jupyterlab/lib/notebook/plugin').notebookHandlerExtension,
     require('jupyterlab/lib/running/plugin').runningSessionsExtension,
     require('jupyterlab/lib/shortcuts/plugin').shortcutsExtension,
